delete edit fix

+++ frontend/server/api/article.delete.ts
@@ -10,16 +10,8 @@ export default defineEventHandler(async (event) => {
 
     const storage = useFileStorage()
 
-    if (!session || session.role !== 'ADMIN') {
-        console.log(`403 admin panel (user: ${session?.id || 'none'})`)
-        throw createError({
-            statusCode: 403,
-            statusMessage: 'Unauthenticated'
-        })
-    }
-
     if (!query.articleId) {
-        console.log(`400 admin panel (user: ${session.id})`)
+        console.log(`400 delete article (user: ${session.id})`)
         throw createError({
             statusCode: 400,
             statusMessage: 'Invalid payload'
@@ -27,6 +19,28 @@ export default defineEventHandler(async (event) => {
     }
 
     const prisma = usePrisma()
+
+    const article = await prisma.article.findUnique({
+        where: {
+            id: Number.parseInt(query.articleId)
+        }
+    })
+
+    if (!article) {
+        console.log(`404 delete article (article: ${query.articleId}) (user: ${session?.id || 'none'})`)
+        throw createError({
+            statusCode: 404,
+            statusMessage: 'Article not found'
+        })
+    }
+
+    if (!(session && (session.role === 'ADMIN' || session.id === article.authorId))) {
+        console.log(`403 delete article (user: ${session?.id || 'none'})`)
+        throw createError({
+            statusCode: 403,
+            statusMessage: 'Unauthenticated'
+        })
+    }
 
     const deletedArticle = await prisma.article.delete({
         where: {
@@ -37,9 +51,9 @@ export default defineEventHandler(async (event) => {
     storage.removeItem(`${deletedArticle.path}/content.md`).catch((err) => console.log(`could not remove content ${deletedArticle.path} (${err})`))
     storage.removeItem(`${deletedArticle.path}/para.json`).catch((err) => console.log(`could not remove para ${deletedArticle.path} (${err})`))
 
-    fs.rmdir(`./userfiles/${deletedArticle.path}`, (err) => console.log(`could not remove dir ${deletedArticle.path} (${err})`))
+    fs.rmdir(`./userfiles/${deletedArticle.path}`, (err) => console.log(`try to remove dir ${deletedArticle.path} (err: ${err})`))
 
-    console.log(`delete article ${deletedArticle.id} by ${session.id}`)
+    console.log(`delete article ${deletedArticle.id} by ${session.id} (role: ${session.role})`)
 
     return deletedArticle
